Notification Manager new_comment fixed. Underlying the move to removing the method to fetch users on a comment thread from the User model is a deeper philosophy of moving in due time to a service oriented architecture.
This requires individual services
to have little knowledge of each other at the controller level and no
knowledge of each other at the model and db level. Thus. all queries
using joins are to be replaced with two queries in two different tables
performed outside the model.

These tiny changes will keep the code modular and enable a smoother
move to a SOA when the time arrives.