Start of the Sweepers construct. This follows a two-fold philosophy.
Model observers are getting crowded with cache management code
which is better placed in a sweeper dedicated to cache management.

Currently the caching only wraps around DB queries whereas profiling
shows that majority of the time in requests is spent generating the json.
By wrapping the to be generated json into a cache_fragment tag i.e. serving
the json directly from memcached without a query or view generation time shall
be a tremendous gain.