Related stores route with caching support
Explanation of three decision.

Every processed store and category combination is stored in specialty
instead of simply the 'top' category. This enables the rake task to
be run automatically simply updating the 'weight' of each specialty.

An is_top column is added to every specialty. Because of the previous
point, to retrieve the 'top' category for a store requires a sort
query on the weight column given a particular store id. This sort query
although fast shall run every time related stores are fetched for a store.
To turn this into a simpler select query an is_top column is added to
every specialty which is updated by the rake task.

It is interesting to note that the store for which related store are being
fetched is not removed from the related stores json. Removing the store in
the query would not only make it slightly slower it'll also relegate caching to
be store level not category level - 100 cache entries vs 8 cache entries. Thus
the onus to remove the current stores from the related stores is on the client.
Not too hard given the upper limit of 11 related stores.