Extending active record to include a select named_scope.
This is part of a larger optimization plan to only fetch columns
from the database that are being used. Since the contents of a query
are fetched over the network such minor optimizations are likely
to significantly reduce db query time especially for queries on stores
or products that fetch a large amount of data from the db.