Payload modified to work directly with ActiveRecord objects.
Invalid/inconsistent states of ActiveRecord model objects is a concern.
Eg - if a User model object is added to the Queue and processed after
an unknown time the model object extracted from the queue may no longer
be in sync with the database.

To workaround this issue, the Payload tracks ActiveRecord objects
and reloads them from the DB to ensure the freshest possible state.