Modified stores update domains rake tasks. *Philosophy There is a tradeoff between updating all stores together via a single rake task or updating each store individually via the processing queue.
The former while far lighter on the database will operate in a single
thread and result in a slow, bulky execution which relies on finishing
fully & correctly for a complete implementation which makes it susceptible
to crashes and process failures, not to mention memory bloat.

The latter sprays the task over the N running instances of the processing
queue and they individually hammer the database. Even though this leads to
far more queries (linear with number of stores) it's preferrable because it
spreads fault tolerance across different machines and processes. This is
especially useful when there are requests to third party services involved and
almost necessary when the number of third party requests are linear with the
number of objects processed (stores in this case).