update top shoppers moved into mailman.
There is a fundamental question about the responsibility of mailman.
How much logic of the application should mailman handle and what
kind of logic.

Controlling the generation of top shoppers seems far too much
for the mailman to execute and violates abstractions on many levels.

It's barely OK for the time being and setting up the pipline with
cron jobs will bring more perspective.

On a separate note it looks like the Email table should be populated
at the Mailman level and not via the UserMailer. More on both these
issues later.