clean solution for updating top shoppers.
This is a far cleaner approach than the one discussed in the previous commit.
Store model becomes responsible for initiating the update process and delegates
updating individual top shoppers to an instance method while providing outside
methods access to the top shoppers for each stores being updated.

Eg - Store.update_top_shoppers{|store,users| Mailman.notify_top_shoppers(store,users)}

This structure respects all the levels of abstraction. Interestingly, this
along with the fact that it'll be called via a cron job opens up the room
for a CronWork class that wakes up and fires the above call.