HealthReports
The introduction of Cron results in more complex execution pipelines
and error scenarios. Although the implementation is abstracted and hence
modular with exception logging at each level there are still failure
cases that can't be tracked efficiently.

For eg - cron server failing, bug in timestamps of cron scheduling,
crashes in corn launching etc.

To cover these scenario a HealthReports sytem is being introduced. The
utility is mainly for crons but applies more generally. By generating a
health report at the end of an execution we can track with certainity whether
a CronWorker has successfully executed. This combined with LoggedExceptions
minimizes the chance of errors sneaking by.