removing sequence based deploy tasks from install.
This starts the idea that code releases must not coincide with
booting up new machines. The code should always be released with migrations
and asset deployments onto the existing machines.

This philosophy allows us to transition towards self provisioning instances
that bootup and install a copy of the codebase and turn on the required
processed automatically. Which also moves us one step closer to auto-scaling.

These ideas all rely on separating releases and installations.

I realize there is a school of thought that suggests new machines should be booted
up for every new release but given the way rails migrations work it seems almost
impossible to pull off while maintaining ACID for user's currently on the site.

For the time being we stick to separating releases and installations.