secondary processing queue.
The secondary queue enables a priority based separation of asynchronous
tasks while maintaining the existing processor daemon and server
architecture.

The same processing daemon handles the secondary queue but any entry in
the primary queue takes priority.

As an example, when products are crawled they need to be processed in bulk
and they are usually hundreds of thousands of tasks added into the queue.
With the secondary queue the same processing servers can handle this
extra volume while giving priority to any critical tasks like mining a user's
friends or publishing a purchase.